small bugfix in dtype of index array
I ran into errors downstream as the row index array was of float dtype. Setting it to int dtype solved the problem

+++ coupling/coupling_PCR_FM/coupling_functions.py
@@ -572,7 +572,7 @@ def getPCRIndices(index_hydro, lat_hydro, lon_hydro, pcrLandMask, mapAttributes,
 	selectedLonPCR= selectedLonPCR[npLandMask == 1]
 	#-create arrays to hold information derived from landmask
 	minimumDistance= np.zeros((index_hydro.size))-1.0
-	rowIndexPCR= np.zeros((index_hydro.size))-1
+	rowIndexPCR= np.zeros((index_hydro.size), dtype=int)-1
 	colIndexPCR= np.zeros((index_hydro.size), dtype= int)-1
 	cellIndexPCR= np.zeros((index_hydro.size), dtype= int)-1
 	#-iterate over all DFM indices and find minimum distance and corresponding
